requirements.txt
pandas==2.1.3
transformers==4.35.2
lightning==2.1.2
tokenizers==0.15.0
tqdm==4.66.1
scikit-learn==1.3.2
numpy==1.24.1
torchmetrics==1.2.1
wandb==0.16.5
matplotlib==3.8.4
rich==13.7.1
biopython==1.83